requirements.txt
absl-py==0.9.0
asgiref==3.3.4
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.2.4
gast==0.2.2
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mysqlclient==2.0.3
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.4.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
